1. deprecated  FeatureInstaller.addMainMenuPlugin() methods 2. added method javadoc
the more generic FeatureInstaller.addMainMenuPlugin() methods are in trunk now for a while and we should promote the use of them. back to one method to rule them all. they also had some issues:
- most only supported AbstractUiPlugIn, which not all plugins are based on
- one exposed adding of an existing MenuItem and a PlugIn

the correct way to add and modify (Popup/Main)MenuItem is to call the appropriate add method and modify the returned MenuItem.

of they are left there.. just marked deprecated so future implementers will hopefully use the new method instead.